# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amap POI/AOI Collector",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tasks/preview": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Preview Task",
        "operationId": "preview_task_api_tasks_preview_post",
        "parameters": [
          {
            "name": "max_rows",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Max Rows"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_task_api_tasks_preview_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskView"
                  },
                  "type": "array",
                  "title": "Response List Tasks Api Tasks Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_tasks_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_task_api_tasks_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/run": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Run Task",
        "operationId": "run_task_api_tasks__task_id__run_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "sync",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Sync"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/aoi": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Run Aoi",
        "operationId": "run_aoi_api_tasks__task_id__aoi_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "sync",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Sync"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AoiRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/pause": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Pause Task",
        "operationId": "pause_task_api_tasks__task_id__pause_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/resume": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Resume Task",
        "operationId": "resume_task_api_tasks__task_id__resume_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/export": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Export Task",
        "operationId": "export_task_api_tasks__task_id__export_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(csv|geojson)$",
              "default": "csv",
              "title": "Format"
            }
          },
          {
            "name": "crs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(gcj02|wgs84)$",
              "default": "wgs84",
              "title": "Crs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AoiRequest": {
        "properties": {
          "record_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Record Ids"
          }
        },
        "type": "object",
        "required": [
          "record_ids"
        ],
        "title": "AoiRequest"
      },
      "Body_create_task_api_tasks_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "name_column": {
            "type": "string",
            "title": "Name Column"
          },
          "adname_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adname Column"
          }
        },
        "type": "object",
        "required": [
          "file",
          "name_column"
        ],
        "title": "Body_create_task_api_tasks_post"
      },
      "Body_preview_task_api_tasks_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_preview_task_api_tasks_preview_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PreviewResult": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "rows": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Rows"
          },
          "suggested_name_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Name Column"
          },
          "suggested_adname_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Adname Column"
          },
          "name_confidence": {
            "type": "string",
            "title": "Name Confidence"
          },
          "row_count_preview": {
            "type": "integer",
            "title": "Row Count Preview"
          }
        },
        "type": "object",
        "required": [
          "columns",
          "rows",
          "suggested_name_column",
          "suggested_adname_column",
          "name_confidence",
          "row_count_preview"
        ],
        "title": "PreviewResult"
      },
...
      "RunResult": {
        "properties": {
          "started": {
            "type": "boolean",
            "title": "Started"
          },
          "pending": {
            "type": "integer",
            "title": "Pending"
          }
        },
        "type": "object",
        "required": [
          "started",
          "pending"
        ],
        "title": "RunResult"
      },
      "TaskCreated": {
        "properties": {
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "total",
          "skipped"
        ],
        "title": "TaskCreated"
      },
      "TaskView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source_filename": {
            "type": "string",
            "title": "Source Filename"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "done": {
            "type": "integer",
            "title": "Done"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "paused_reason": {
            "type": "string",
            "title": "Paused Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "source_filename",
          "total",
          "done",
          "failed",
          "paused_reason"
        ],
        "title": "TaskView"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}